requirements.txt
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.1.2
psycopg2-binary==2.8.5
ipython==7.13.0